file-converter/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# Google Cloud
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-api-python-client==2.134.0

# Conversion libraries
markdown==3.6
reportlab==4.2.2
python-docx==1.1.2
openpyxl==3.1.4
